requirements.txt
xrpl-py>=2.0.0